Final Commit after the Presentation
