+++ force-app/main/default/classes/ContactHandler.cls
@@ -13,7 +13,7 @@ public with sharing class ContactHandler {
 
     @AuraEnabled
     public static Contact createCandidate(String firstName, String lastName, String email){
-        Contact newContact = new Contact(FirstName = firstName, LastName = lastName, Email = email);
+        Contact newContact = new Contact(FirstName = firstName, LastName = lastName, Email = email, Type__c = 'Job Applicant');
         insert newContact;
         return newContact;
     }
