+++ force-app/main/default/labels/CustomLabels.labels-meta.xml
@@ -7,101 +7,4 @@
         <shortDescription>Duplicate_Event_Error</shortDescription>
         <value>The event start/end date/time is conflicting with another scheduled meeting.  Please correct start/end date/time and try</value>
     </labels>
-    <labels>
-        <fullName>Opp_Amount_More_Than_5000</fullName>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>Opp Amount More Than 5000</shortDescription>
-        <value>Opportunity amount must be greater than 5000</value>
-    </labels>
-    <labels>
-        <fullName>Prevent_Opportunity_Deletion_For_Banking</fullName>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>Prevent Opportunity Deletion For Banking</shortDescription>
-        <value>Cannot delete closed opportunity for a banking account that is won</value>
-    </labels>
-    <labels>
-        <fullName>fflib_QueryFactory_crossobject_fieldsets_not_allowed_error</fullName>
-        <language>en_US</language>
-        <protected>false</protected>
-        <shortDescription>Error when selecting a cross-object fieldsset is not allowed and attempted.</shortDescription>
-        <value>Cross-object fields not allowed and field &quot;{0}&quot; is a cross-object field.</value>
-    </labels>
-    <labels>
-        <fullName>fflib_QueryFactory_fieldset_wrong_sobject_error</fullName>
-        <language>en_US</language>
-        <protected>false</protected>
-        <shortDescription>Error when selecting a field set for a different sobject type.</shortDescription>
-        <value>Field set &quot;{0}&quot; is not for SObject type &quot;{1}&quot;</value>
-    </labels>
-    <labels>
-        <fullName>fflib_QueryFactory_recursive_subquery_error</fullName>
-        <language>en_US</language>
-        <protected>false</protected>
-        <shortDescription>Presented when attempting to nest subqueries.</shortDescription>
-        <value>Invalid call to subselectQuery.  You may not add a subselect query to a subselect query.</value>
-    </labels>
-    <labels>
-        <fullName>fflib_QueryFactory_subquery_invalid_relationship</fullName>
-        <language>en_US</language>
-        <protected>false</protected>
-        <shortDescription>Presented when attempting to create a subquery from an invalid relationship.</shortDescription>
-        <value>Invalid call to subselectQuery.  Invalid relationship for table.</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_field_not_insertable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_field_not_insertable</shortDescription>
-        <value>You do not have permission to insert field {1} on {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_field_not_readable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_field_not_readable</shortDescription>
-        <value>You do not have permission to read the field {1} on {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_field_not_updateable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_field_not_updateable</shortDescription>
-        <value>You do not have permission to update the field {1} on {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_object_not_deletable</fullName>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_object_not_deletable</shortDescription>
-        <value>You do not have permission to delete {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_object_not_insertable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_object_not_insertable</shortDescription>
-        <value>You do not have permission to insert {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_object_not_readable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_object_not_readable</shortDescription>
-        <value>You do not have permission to read {0}</value>
-    </labels>
-    <labels>
-        <fullName>fflib_security_error_object_not_updateable</fullName>
-        <categories>security,error</categories>
-        <language>en_US</language>
-        <protected>true</protected>
-        <shortDescription>fflib_security_error_object_not_updateable</shortDescription>
-        <value>You do not have permission to update {0}</value>
-    </labels>
 </CustomLabels>
