+++ force-app/main/default/lwc/jobListing/jobListing.js-meta.xml
@@ -4,5 +4,7 @@
     <isExposed>true</isExposed>
     <targets>
         <target>lightning__AppPage</target>
+         <target>lightning__RecordPage</target>
+    <target>lightning__HomePage</target>
     </targets>
 </LightningComponentBundle>
